Change NetList writer paradigm.
Give more control to users : they can decide not to use the DSL to write
the AST manually, allowing them to set the name for the variable
manually. The DSL is still here, and can be used as a convenience to
write some subfunctions, but the whole circuit is not expected to be in
the DSL anymore.